refactor: 에산 관련 상수 이동 및 적용 #135
