+++ fe/src/components/budgets/steps/TemplateSelectionStep.tsx
@@ -1,6 +1,6 @@
 import { CheckCircle2, Info } from 'lucide-react';
 
-import { BUDGET_TEMPLATES } from '@/constants/budgetTemplates';
+import { BUDGET_TEMPLATES } from '@/constants/budget';
 import { Card } from '@/components/ui/card';
 import DialogStepHeader from '@/components/budgets/steps/DialogStepHeader';
 import { TemplateId } from '@/types/budgetGuide';
